排行榜修改

+++ src/main/java/tw/ispan/librarysystem/controller/comment/BookCommentController.java
@@ -93,6 +93,6 @@ public class BookCommentController {
         if (member == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
         }
-        return ResponseEntity.ok(member.getId());
+        return ResponseEntity.ok(member.getId().intValue());
     }
 }
